requirements.txt
aiomysql==0.2.0
PyMySQL==1.1.1
python-dotenv==1.1.1